Resolve 'branch-killing' issue
CloudSQL can't use the same instance name until >1 week after deletion.
If you create an instance, you need to be able to connect to it.
If a pipeline fails after an instance is created, the artifacts stage never happens so the root password isn't captured for the next pipeline.
This change sends the file direct to GCS so it's available for future pipelines even if the one that creates the cloudSQL instance fails.